# application/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Mechanic API
  description: |
    A comprehensive REST API for managing auto repair shop operations including mechanics, customers, inventory, and service tickets.
    
    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. After logging in, include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - Customer endpoints: 100 requests per hour
    - Login endpoints: 10 requests per minute
    - Specific endpoints may have additional rate limits
    
    ## Error Handling
    The API uses standard HTTP status codes and returns error messages in JSON format.
  version: 1.0.0
  contact:
    name: API Support
    email: support@mechanicapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.mechanic.com/v1
    description: Production server
...
paths:
...
  /mechanic/register:
    post:
      tags:
        - Mechanics
      summary: Register a new mechanic
      description: Create a new mechanic account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MechanicRegistration'
      responses:
        '201':
          description: Mechanic registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MechanicResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mechanic/login:
    post:
      tags:
        - Mechanics
      summary: Mechanic login
      description: Authenticate mechanic and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: mechanic@example.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  mechanic:
                    $ref: '#/components/schemas/MechanicResponse'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mechanic/stats:
    get:
      tags:
        - Mechanics
      summary: Get mechanics with ticket counts
      description: Retrieve all mechanics with their associated service ticket counts
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    mechanic:
                      $ref: '#/components/schemas/MechanicResponse'
                    ticket_count:
                      type: integer
                      example: 5
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mechanic/cached:
    get:
      tags:
        - Mechanics
      summary: Get cached mechanics list
      description: Retrieve mechanics list with caching (60-second cache)
      responses:
        '200':
          description: Mechanics retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MechanicResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mechanic/rate-limited:
    get:
      tags:
        - Mechanics
      summary: Rate limited test endpoint
      description: Test endpoint with rate limiting (10 requests per minute)
      responses:
        '200':
          description: Request successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "This route is rate limited"
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mechanic/:
    get:
      tags:
        - Mechanics
      summary: Get all mechanics
      description: Retrieve all mechanics (requires authentication)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Mechanics retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MechanicResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mechanic/profile:
    get:
      tags:
        - Mechanics
      summary: Get current mechanic profile
      description: Retrieve the profile of the authenticated mechanic
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MechanicResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Mechanic not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Mechanics
      summary: Update current mechanic profile
      description: Update the profile of the authenticated mechanic
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MechanicUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MechanicResponse'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Mechanic not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mechanic/{mechanic_id}:
    get:
      tags:
        - Mechanics
      summary: Get mechanic by ID
      description: Retrieve a specific mechanic by their ID
      security:
        - BearerAuth: []
      parameters:
        - name: mechanic_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Mechanic retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MechanicResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Mechanic not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Mechanics
      summary: Update mechanic by ID
      description: Update a specific mechanic by their ID
      security:
        - BearerAuth: []
      parameters:
        - name: mechanic_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MechanicUpdate'
      responses:
        '200':
          description: Mechanic updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MechanicResponse'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Mechanic not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Mechanics
      summary: Delete mechanic by ID
      description: Delete a specific mechanic by their ID
      security:
        - BearerAuth: []
      parameters:
        - name: mechanic_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Mechanic deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Mechanic deleted successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Mechanic not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Customer Routes
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoints
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Registration failed"
        message:
          type: string
          example: "Internal server error occurred"
        details:
          type: string
          example: "Database connection error"
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Not a valid email address"]
            password: ["Password must be at least 8 characters"]

    # Authentication Schemas
...
    MechanicResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        specialization:
          type: string
          example: "Engine Repair"
        phone:
          type: string
          example: "123-456-7890"
        address:
          type: string
          example: "123 Main St, City, State"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    MechanicRegistration:
      type: object
      required:
        - name
        - email
        - password
        - specialization
      properties:
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          example: "securepassword123"
        specialization:
          type: string
          example: "Engine Repair"
        phone:
          type: string
          example: "123-456-7890"
        address:
          type: string
          example: "123 Main St, City, State"
    MechanicUpdate:
      type: object
      properties:
        name:
          type: string
          example: "John Smith"
        email:
          type: string
          format: email
          example: "john.smith@example.com"
        specialization:
          type: string
          example: "Transmission Repair"
        phone:
          type: string
          example: "987-654-3210"
        address:
          type: string
          example: "456 Oak Ave, City, State"
        password:
          type: string
          example: "newpassword123"

    # Customer Schemas
...